Created playerProfile folder and moved PlayerProfile component there
