+++ view/src/components/playerProfile/PlayerProfile.js
